+++ src/test/java/matcheam/match/MatchRepositoryTest.java
@@ -1,21 +1,20 @@
 package matcheam.match;
 
 import com.ninja_squad.dbsetup.DbSetup;
-import com.ninja_squad.dbsetup.operation.DeleteAll;
 import com.ninja_squad.dbsetup.operation.Operation;
 import matcheam.support.TestContext;
+import org.assertj.core.api.SoftAssertions;
 import org.junit.Before;
 import org.junit.Test;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
-import java.util.Collection;
 import java.util.List;
 
 import static com.ninja_squad.dbsetup.Operations.deleteAllFrom;
 import static com.ninja_squad.dbsetup.Operations.insertInto;
 import static com.ninja_squad.dbsetup.Operations.sequenceOf;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.Assert.*;
 
 /**
  * Created by ooguro on 2017/07/08.
@@ -98,6 +97,32 @@ public class MatchRepositoryTest {
 	public void 指定したIDの募集が存在しない場合nullが返ってくること() throws Exception {
 		Match actual = matchRepository.findBy(new Identifier("X"));
 		assertThat(actual).isNull();
+	}
+	@Test
+	public void 募集内容を登録できること() throws Exception {
+		Match match = getRecord();
+
+		Identifier identifier = matchRepository.register(match);
+
+		Match actual = matchRepository.findBy(identifier);
+		SoftAssertions softAssertions = new SoftAssertions();
+		softAssertions.assertThat(actual.getPlace()).isEqualTo("場所");
+		softAssertions.assertThat(actual.getDate()).isEqualTo(LocalDate.of(2017, 1, 25));
+		softAssertions.assertThat(actual.getTime()).isEqualTo("2時間");
+		softAssertions.assertThat(actual.getLevel()).isEqualTo(Level.LEVEL1);
+		softAssertions.assertThat(actual.getStart()).isEqualTo("昼ぐらい");
+		softAssertions.assertAll();
+	}
+
+	private Match getRecord() {
+		Match match = new Match();
+		match.setPlace("場所");
+		match.setStart("昼ぐらい");
+		match.setDate(LocalDate.of(2017, 1, 25));
+		match.setTime("2時間");
+		match.setLevel(Level.LEVEL1);
+		match.setMaxPlayers(BigDecimal.TEN);
+		return match;
 	}
 }
